Update to TypeScript
